+++ examples/comet-cygnus/04.create-context-data.sh
@@ -15,7 +15,7 @@ fi
 set -eu
 for i in {0..9}
 do
-  echo $i
+  echo "$i"
   ngsi upsert \
     --host "${ORION}" \
     --service openiot \