+++ setup/clean.sh
@@ -1,9 +1,20 @@
 #!/bin/sh
+FORCE_CLEAN=${FORCE_CLEAN:-false}
+
+if [ "${FORCE_CLEAN}" != "true" ]; then
+  echo "!CAUTION! This command cleans up your FIWARE instance including your all data."
+  read -p "Are you sure you want to run it? (Y/n): " ans
+
+  if [ "${ans}" != "Y" ]; then
+    echo "Canceled"
+    exit 1
+  fi
+fi
 
 for file in docker-compose.yml docker-cert.yml docker-idm.yml
 do
   if [ -e ${file} ]; then
-    sudo /usr/local/bin/docker-compose -f ${file} down --remove-orphan
+    /usr/bin/docker compose -f ${file} down --remove-orphans
     sleep 5
     rm -f ${file}
   fi
@@ -11,6 +22,10 @@ done
 
 if [ -d ./data ]; then
   sudo rm -fr data
+fi
+
+if [ -e setup_ngsi_go.sh ]; then
+  sudo rm -f setup_ngsi_go.sh
 fi
 
 sudo rm -fr ./config